+++ tests/Function/Trait/SqlWalkerTestTrait.php
@@ -17,7 +17,7 @@ use Mockery\MockInterface;
 
 trait SqlWalkerTestTrait
 {
-    private function createMysqlSqlWalker(): SqlWalker|MockInterface
+    protected function createMysqlSqlWalker(): SqlWalker|MockInterface
     {
         $connection = Mockery::mock(Connection::class);
         $connection->shouldReceive('getDatabasePlatform')
@@ -30,7 +30,7 @@ trait SqlWalkerTestTrait
         return $sqlWalker;
     }
 
-    private function createNonMysqlSqlWalker(): SqlWalker|MockInterface
+    protected function createNonMysqlSqlWalker(): SqlWalker|MockInterface
     {
         $connection = Mockery::mock(Connection::class);
         $connection->shouldReceive('getDatabasePlatform')